MM-46453,MM-32912: fixes 26 long channel names
Currently we support both id, and name in the URL for channels. The
distinguish between the 2 relies on the length of the string, i.e if we
encounter an unknown channel name and the length of the name is 26 chars
it must be an ID, which causes errors on public channels with 26 chars
in the name.

This PR tries to solve that by fetching the channel by id when it hits
the above case, so,
- if the response is OK (2**) then it's an id,
- if the response is 404 then it's a channel_name,
- in any other error case it returns error.