MM-46634: fixes opening a notification
When you have the app minimized and receive messages clicking a
notification will focus the window and then redirect to the channel for
which the notification was clicked.
If the current channel has unread messages focusing the window and then
redirecting would mark current channel as read, which is unintended.

This PR fixes that issue, by first redirecting and then focusing the
window.