MM-46410: adds urgency on mention counts
We have introduced priority for posts in
mattermost/mattermost-webapp#10951. We do need to color the mention badges
in the webapp with a prominent color when a mention is posted in an urgent message.
A thread has urgent mentions if the root post is marked as urgent,
and the replies contain mentions to the user viewing the thread.

This PR incorporates the newly added `total_unread_urgent_mentions`, and
`urgent_mention_count` properties in the responese for thread counts,
and channel members respectively.
The team's icon in the sidebar, threads link, and channel sidebar link,
are all shown with an urgent mention count (red background) when they
contain an mention in an urgent post.

unreadMentionDiff