Fixes duplicated threads in thread list
Upon leaving a channel we collect all threads that belong to that
channel and then we remove them from the threadsInTeam lists.
BUT if a thread is in threads, but not in threadsInTeam then depending
on the position of that thread, it will duplicate the list from that
position onwards.
This is happening because we didn't check index against -1.
So the slice would be [...list.slice(0, -1), ...list.slice(0)],
which produces "crazy" results.

This commit fixes that issue by checking each thread's index in
threadsInTeam list, and if it doesn't exist it does nothing.