MM-48565: fixes delay in channel urgent mention
When we mark a channel as unread through a post which has mentions in
urgent posts, we show briefly the normal badge, and then we show the
"urgent" badge.

It seems that we didn't pass the urgent mention counts upon the websocket
event handling for post_unread_event.
This commit fixes that and it shows the urgent badge immediately.