Handle persistent desktop notifications
When a persistent notification is triggered
(ws 'persistent_notification_triggered') we need to send a desktop
notification, which overrides channel members notify props, to all
mentioned users.

This commit handles that case by bypassing some checks when
'sendDesktopNotification' receives the 'persistent' argument as true.
Which is only set to true when a 'persistent_notification_triggered'
event is received.