MM-49487: fixes post priority picker in safari
For some reason when, Boards Product is enabled, and the .Menu CSS class
is overwritten by it, the ordering of classes in Safari is reversed than
in Firefox which makes the post priority picker not to be visible.

This commit fixes that by bumping the styled components rule's
specificity.

See https://styled-components.com/docs/faqs#how-can-i-override-styles-with-higher-specificity