Fix spurious redirect on playbooks upgrade (#12022)
For the past few months, when Playbooks is upgraded on community, some users get redirected to the `/error?type=team_not_found` page.  We've been able to reproduce it locally when already on `/playbooks`, but mysteriously it also happens on community wherever you are.

The timing and code changes appear to date back to https://mattermost.atlassian.net/browse/MM-43306 and logic therein that fails to take into account reserved team names. Change this to ignore such transitions, with the hope it addresses the issue on community as well.

Co-authored-by: Mattermost Build <build@mattermost.com>