MM-49088: fixes file uploading state
File uploading should not update the draft in a debounced fashion.
It should update immediately and should not be interrupted by message
changes.

This commit was authored to fix a bug where file uploads appear as
processing for ever if a user starts typing while the upload is in
processing state.